Move some code to prevent duplicated imports of *.spec.ts files. They should never be imported!
